Main Motion Control Software
This code was in made in CRBasic following the MATLAB cartesian vector simulations. Since CRBasic has different syntax some of the flow of the code may not be as similar as in MATLAB, but it still follows the same principal. Note that the software does not follow an official method for motion control and waypoint tracking, as this is a preliminary code for testing.